net: wireless: bcmdhd: fix hang in wifi_get_cached_gscan_results
Bug: 15144213

(cherry picked from commit 5d96ba26b7665ca57e50d22de54bb8a0a6a40039)

+++ drivers/net/wireless/bcmdhd/wl_cfgvendor.c
@@ -235,8 +235,8 @@ static int wl_cfgvendor_gscan_get_batch_results(struct wiphy *wiphy,
 	             (num_scan_ids * GSCAN_BATCH_RESULT_HDR_LEN) +
 	             VENDOR_REPLY_OVERHEAD + SCAN_RESULTS_COMPLETE_FLAG_LEN;
 
-	if (mem_needed > NLMSG_DEFAULT_SIZE) {
-		mem_needed = NLMSG_DEFAULT_SIZE;
+	if (mem_needed > (int32)NLMSG_DEFAULT_SIZE) {
+		mem_needed = (int32)NLMSG_DEFAULT_SIZE;
 		complete = 0;
 	} else {
 		complete = 1;
